uvm_generator_20_01_26/uvm_gen_pipeline/requirements.txt
# UVM Test Case Generator Dependencies
openai>=1.12.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
rich>=13.7.0
click>=8.1.7
jinja2>=3.1.2
